backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9

pydantic>=2.8
pydantic-settings>=2.4

sqlalchemy>=2.0
aiosqlite>=0.20
psycopg[binary]>=3.2

numpy>=1.26
faiss-cpu>=1.8.0

google-genai>=1.0.0
tenacity>=8.3
httpx>=0.27

# Local embeddings fallback (used if Gemini embeddings 404 for your key)
fastembed>=0.6.0

pypdf>=5.0
python-docx>=1.1

orjson>=3.10
cachetools>=5.5
python-dotenv>=1.0
